Turret animation and bank UI
I added recoil to the turret node when it shoots, as well as a damage animation when a projectile strikes an enemy. I also created the beginning UI for a bank that holds a label containing the amount of gold the player has. The player starts with 150 gold, loses 100 for each turret placed, and gains 15 for every enemy defeated.